move _chartobject.py to _buider.py
