+++ bokeh/charts/area.py
@@ -26,7 +26,7 @@ except ImportError:
     print("bokeh.charts needs numpy installed to work properly!")
     raise
 
-from ._chartobject import Builder, create_and_build
+from ._builder import Builder, create_and_build
 from ..models import ColumnDataSource, DataRange1d, GlyphRenderer, Range1d
 from ..models import Patch
