+++ bokeh/charts/bar.py
@@ -25,7 +25,7 @@ except ImportError:
     print("bokeh.charts.Bar needs numpy installed to work properly!")
     raise
 
-from ._chartobject import Builder, create_and_build
+from ._builder import Builder, create_and_build
 from ..models import ColumnDataSource, FactorRange, GlyphRenderer, Range1d
 from ..models.glyphs import Rect
