+++ bokeh/charts/catheatmap.py
@@ -17,7 +17,7 @@ the arguments to the Chart class and calling the proper functions.
 #-----------------------------------------------------------------------------
 from __future__ import print_function, division
 
-from ._chartobject import Builder, create_and_build
+from ._builder import Builder, create_and_build
 from ._data_adapter import DataAdapter
 from ..models import ColumnDataSource, FactorRange, GlyphRenderer, HoverTool
 from ..models.glyphs import Rect
