+++ bokeh/charts/dot.py
@@ -22,7 +22,7 @@ try:
 except ImportError:
     pd = None
 
-from ._chartobject import Builder, create_and_build
+from ._builder import Builder, create_and_build
 from ..models import ColumnDataSource, FactorRange, GlyphRenderer, Range1d
 from ..models.glyphs import Segment
