+++ bokeh/charts/scatter.py
@@ -25,7 +25,7 @@ except:
     pd = None
 
 from collections import OrderedDict
-from ._chartobject import create_and_build, Builder
+from ._builder import create_and_build, Builder
 from ._data_adapter import DataAdapter
 from ..models import ColumnDataSource, Range1d
