+++ bokeh/charts/timeseries.py
@@ -23,7 +23,7 @@ try:
 except ImportError:
     pd = None
 
-from ._chartobject import Builder, create_and_build
+from ._builder import Builder, create_and_build
 from ..models import ColumnDataSource, DataRange1d, GlyphRenderer, Range1d
 from ..models.glyphs import Line
